# packages/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EVS API",
    "description": "USACE Enterprise Visibility Suite API.\n\nEvery endpoint that replaces an Oracle APEX page carries an `x-apex-page` extension so the\ntraceability matrix (legacy/traceability.csv) can be regenerated from the contract.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workforce/labor": {
      "get": {
        "tags": [
          "workforce"
        ],
        "summary": "EMS labor by district and pay period (APEX People IR)",
        "description": "Interactive Report parameters as on /projects. Columns: district, pay_period, hours_plan, hours_regular, hours_overtime, labor_cost.",
        "operationId": "labor_api_v1_workforce_labor_get",
        "parameters": [
          {
            "name": "fiscal_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2026,
              "title": "Fiscal Year"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "name": "pay_period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pay Period"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma separated columns, `-` prefix for descending",
              "title": "Sort"
            },
            "description": "Comma separated columns, `-` prefix for descending"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text search across text columns (IR search bar)",
              "title": "Q"
            },
            "description": "Free text search across text columns (IR search bar)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv)$",
              "description": "`csv` streams the filtered rows",
              "default": "json",
              "title": "Format"
            },
            "description": "`csv` streams the filtered rows"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaborSummary"
                }
              },
              "text/csv": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-page": "40",
        "x-apex-authorization": "Authenticated User"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AsOf": {
        "properties": {
          "source_as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source As Of",
            "description": "Newest timestamp inside the source payload"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At",
            "description": "When the EVS ingestion worker stored it"
          },
          "freshness": {
            "type": "string",
            "enum": [
              "fresh",
              "aging",
              "stale",
              "simulated"
            ],
            "title": "Freshness",
            "default": "fresh"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "live | fixtures | simulated | synthetic",
            "default": "fixtures"
          }
        },
        "type": "object",
        "title": "AsOf",
        "description": "Every dataset carries both the source timestamp and when EVS fetched it."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LaborRow": {
        "properties": {
          "district": {
            "type": "string",
            "title": "District"
          },
          "pay_period": {
            "type": "string",
            "title": "Pay Period"
          },
          "hours_plan": {
            "type": "number",
            "title": "Hours Plan"
          },
          "hours_regular": {
            "type": "number",
            "title": "Hours Regular"
          },
          "hours_overtime": {
            "type": "number",
            "title": "Hours Overtime"
          },
          "labor_cost": {
            "type": "number",
            "title": "Labor Cost"
          }
        },
        "type": "object",
        "required": [
          "district",
          "pay_period",
          "hours_plan",
          "hours_regular",
          "hours_overtime",
          "labor_cost"
        ],
        "title": "LaborRow"
      },
      "LaborSummary": {
        "properties": {
          "fiscal_year": {
            "type": "integer",
            "title": "Fiscal Year"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/LaborRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page"
              },
              {
                "type": "null"
              }
            ]
          },
          "as_of": {
            "$ref": "#/components/schemas/AsOf"
          }
        },
        "type": "object",
        "required": [
          "fiscal_year",
          "rows",
          "as_of"
        ],
        "title": "LaborSummary"
      },
...
      "Page": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset"
        ],
        "title": "Page"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}